even more clean up
